api/requirements.txt
azure-functions
pymongo[srv]==4.6.1
dnspython==2.6.1